third_party/WebKit/LayoutTests/http/tests/workers/worker-importScripts-onerror-sameorigin.html

<!DOCTYPE html>
<html>
<head>
    <script>
        window.jsTestIsAsync = true;
        window.isOnErrorTest = true;
    </script>
    <script src="/resources/js-test-pre.js"></script>
    <script src="/resources/js-test-post.js"></script>
</head>
<body>
    <script>
        description("Ensure that scripts imported into a Worker from same-origin hosts trigger detailed onerror messages.");

        var worker = new Worker('resources/worker-importscripts-onerror-sameorigin.js');

        var workerOnerror;
        worker.onmessage = function (e) {
            workerOnerror = e.data;
            shouldBeEqualToString("workerOnerror.message", "Uncaught This is a custom error message.");
            shouldBeEqualToString("workerOnerror.filename", "http://127.0.0.1:8000/workers/resources/worker-importscripts-onerror-sameorigin.js");
            shouldEvaluateTo("workerOnerror.lineno", 5);
            shouldEvaluateTo("workerOnerror.column", 0);
        };

        var onerrorMessage, onerrorURL, onerrorLine;
        worker.onerror = function (e) {
            onerrorMessage = e.message;
            onerrorURL = e.filename;
            onerrorLine = e.lineno;
            shouldBeEqualToString("onerrorMessage", "Uncaught This is a custom error message.");
            shouldBeEqualToString("onerrorURL", "http://127.0.0.1:8000/workers/resources/worker-importscripts-onerror-sameorigin.js");
            shouldEvaluateTo("onerrorLine", 5);

            e.preventDefault();
            finishJSTest();
        };
    </script>
    <script src="/resources/js-test-post.js"></script>
</body>
</html>
